# docs/architecture/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/editorial/posts": {
      "get": {
        "operationId": "EditorialController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EditorialPostDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      },
      "post": {
        "operationId": "EditorialController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdDto"
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/editorial/posts/{id}/assignment": {
      "patch": {
        "operationId": "EditorialController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/editorial/revisions/{id}": {
      "patch": {
        "operationId": "EditorialController_edit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdDto"
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/editorial/revisions/{id}/submit": {
      "post": {
        "operationId": "EditorialController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdDto"
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/editorial/revisions/{id}/approve": {
      "post": {
        "operationId": "EditorialController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdDto"
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/editorial/revisions/{id}/publish": {
      "post": {
        "operationId": "EditorialController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdDto"
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/editorial/revisions/{id}/schedule": {
      "post": {
        "operationId": "EditorialController_schedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdDto"
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/editorial/revisions/{id}/cancel": {
      "post": {
        "operationId": "EditorialController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdDto"
                }
              }
            }
          }
        },
        "tags": [
          "editorial"
        ]
      }
    },
    "/api/posts": {
      "get": {
        "operationId": "PublicContentController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicPostSummaryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "public-content"
        ]
      }
    },
    "/api/posts/{slug}": {
      "get": {
        "operationId": "PublicContentController_get",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostDto"
                }
              }
            }
          }
        },
        "tags": [
          "public-content"
        ]
      }
    },
    "/api/discovery/count": {
      "get": {
        "operationId": "DiscoveryController_count",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SitemapCountDto"
                }
              }
            }
          }
        },
        "tags": [
          "public-content"
        ]
      }
    },
    "/api/discovery/entries": {
      "get": {
        "operationId": "DiscoveryController_entries",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SitemapEntryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "public-content"
        ]
      }
    },
    "/api/identity/access": {
      "get": {
        "operationId": "IdentityController_access",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
    "/api/identity/invitations": {
      "post": {
        "operationId": "IdentityController_invite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
    "/api/identity/invitations/accept": {
      "post": {
        "operationId": "IdentityController_accept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "identity"
        ]
      }
    },
    "/api/bookmarks": {
      "get": {
        "operationId": "BookmarkController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookmarkDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "interactions"
        ]
      }
    },
    "/api/bookmarks/{postId}": {
      "put": {
        "operationId": "BookmarkController_save",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "interactions"
        ]
      },
      "delete": {
        "operationId": "BookmarkController_remove",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "interactions"
        ]
      }
    }
  },
  "info": {
    "title": "Kara API",
    "description": "",
    "version": "1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "RevisionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "scheduledAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "body",
          "status",
          "version",
          "scheduledAt"
        ]
      },
      "EditorialPostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "assignedToId": {
            "type": "string",
            "nullable": true
          },
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevisionDto"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "ownerId",
          "assignedToId",
          "revisions"
        ]
      },
      "CreatePostInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "summary",
          "body",
          "slug"
        ]
      },
      "IdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "AssignInput": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
      "EditInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "title",
          "summary",
          "body",
          "version"
        ]
      },
      "ScheduleInput": {
        "type": "object",
        "properties": {
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "scheduledAt"
        ]
      },
      "PublicPostSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "publishedAt"
        ]
      },
      "PublicPostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "body": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "publishedAt",
          "body"
        ]
      },
      "SitemapCountDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          }
        },
        "required": [
          "count"
        ]
      },
      "SitemapEntryDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "publishedAt"
        ]
      },
      "AccessDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "permissions"
        ]
      },
      "InviteInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "creator",
              "reviewer",
              "publisher",
              "administrator"
            ]
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "invitationUrl": {
            "type": "string"
          }
        },
        "required": [
          "invitationUrl"
        ]
      },
      "AcceptInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "BookmarkDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "postId",
          "slug",
          "title"
        ]
      }
    }
  }
}